requirements.txt
absl-py==1.0.0
asgiref==3.5.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
decorator==4.4.2
Django==2.2.5
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
idna==3.3
imageio==2.19.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.2
joblib==1.1.0
JPype1==1.3.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
konlpy==0.6.0
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
matplotlib==3.5.2
moviepy==1.0.3
mysqlclient==2.1.0
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
proglog==0.1.10
protobuf==3.19.4
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wincertstore==0.2
